+++ src/components/wall/wallLayout/index.ts
@@ -1,3 +1,3 @@
-export * from 'components/wall/wallLayout/AddBlockModal'
-export * from 'components/wall/wallLayout/WallHeader'
-export * from 'components/wall/wallLayout/WallHeaderAction'
+export * from '@/components/wall/wallLayout/AddBlockModal';
+export * from '@/components/wall/wallLayout/wallHeader/WallHeader';
+export * from '@/components/wall/wallLayout/wallHeader/WallHeaderAction';
